Dz1. Part 1 (Add start activity and buttons)
