+++ app/src/main/java/com/gmail/yalovikdima/itacademy/MainActivity.java
@@ -1,72 +1,43 @@
 package com.gmail.yalovikdima.itacademy;
 
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
+import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 
-public class MainActivity extends AppCompatActivity implements View.OnClickListener {
+import com.gmail.yalovikdima.itacademy.dz1.Dz1Activity;
 
-    private Button buttonChange;
-    private TextView textView1;
-    private TextView textView2;
+public class MainActivity extends AppCompatActivity implements View.OnClickListener{
+
+    private Button buttonDz1;
+    private Button buttonDz2;
+
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        buttonChange = findViewById(R.id.button_change);
-        textView1 = findViewById(R.id.textView1);
-        textView2 = findViewById(R.id.textView2);
-
-        buttonChange.setOnClickListener(clickListener);
-        textView1.setOnClickListener(this);
-        textView2.setOnClickListener(new View.OnClickListener(){
-            @Override
-            public void onClick(View view) {
-                changeText();
-                changeColorBtnAndTxtV(textView1);
-            }
-        });
+        buttonDz1 = findViewById(R.id.dz1Button);
+        buttonDz1.setOnClickListener(this);
+        buttonDz2 = findViewById(R.id.dz2Button);
+        buttonDz2.setOnClickListener(this);
 
     }
 
     @Override
     public void onClick(View view) {
-        changeText();
-        changeColorBtnAndTxtV(textView2);
-    }
-
-    private View.OnClickListener clickListener = new View.OnClickListener() {
-        @Override
-        public void onClick(View view) {
-            changeText();
-            changeColorTxtVAndTextV();
+        switch (view.getId()){
+            case R.id.dz1Button:{
+                startActivity(Dz1Activity.getIntent(this));
+                break;
+            }
+            case R.id.dz2Button:{
+                break;
+            }
         }
-    };
-
-
-    private void changeText(){
-        String temp = textView1.getText().toString();
-        textView1.setText(textView2.getText().toString());
-        textView2.setText(temp);
     }
 
-    private void changeColorBtnAndTxtV(TextView textView){
-        Drawable drawable_bnt = buttonChange.getBackground();
-        Drawable drawable_tw1 = textView.getBackground();
-        buttonChange.setBackground(drawable_tw1);
-        textView.setBackground(drawable_bnt);
 
-    }
-
-    private  void changeColorTxtVAndTextV(){
-        Drawable drawable_tw1 = textView1.getBackground();
-        Drawable drawable_tw2 = textView2.getBackground();
-        textView1.setBackground(drawable_tw2);
-        textView2.setBackground(drawable_tw1);
-    }
 }
