+++ app/src/main/res/drawable/ic_logo.xml
@@ -1,0 +1,19 @@
+<vector android:height="24dp" android:viewportHeight="303.3"
+    android:viewportWidth="229.4" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#9D9D9C" android:pathData="M174.7,100.4l0.1,36.9c0,0 16.8,-6.1 38.8,-30.2c7.3,-8 10.7,-18.8 10,-26.4c-1,-10.2 -3.3,-15.9 -15.5,-19C201.5,79.2 174.7,100.4 174.7,100.4z"/>
+    <path android:fillColor="#9D9D9C" android:pathData="M28.7,145.5c-3.6,8.2 -4.3,15.7 -0.5,22.5c2.8,5 7.6,9.9 16,11.6c10.1,2.1 18.9,1.5 18.9,1.5v-32.3c0,0 -11,-0.2 -19.6,-0.2C37.9,148.7 33.5,146.3 28.7,145.5z"/>
+    <path android:fillColor="#36A9E1" android:pathData="M71.3,98.9l0,131.9l42.1,0.2l0.1,-154.3C98.3,83.7 87.9,88.5 71.3,98.9z"/>
+    <path android:fillColor="#36A9E1" android:pathData="M113.2,21.4L113.3,0H71.5v40.3C83.4,34 93,28.4 113.2,21.4z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M4.2,260.4v32.3H0v-32.3H4.2z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M17.6,264H7.8v-3.5h23.9v3.5h-9.9v28.7h-4.2V264z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M42.4,278.2v3.1H30.6v-3.1H42.4z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M52.5,282.5l-3.4,10.1h-4.3l11,-32.3h5l11,32.3h-4.5L64,282.5H52.5zM63.1,279.3l-3.2,-9.3c-0.7,-2.1 -1.2,-4 -1.7,-5.9h-0.1c-0.5,1.9 -1,3.9 -1.6,5.8l-3.2,9.3H63.1z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M92.4,291.8c-1.1,0.6 -3.5,1.3 -6.7,1.3c-7,0 -11.5,-4.7 -11.5,-11.8c0,-7.1 4.9,-12.3 12.4,-12.3c2.5,0 4.7,0.6 5.8,1.2l-1,3.3c-1,-0.6 -2.6,-1.1 -4.9,-1.1c-5.3,0 -8.2,3.9 -8.2,8.8c0,5.4 3.4,8.7 8,8.7c2.4,0 4,-0.6 5.2,-1.1L92.4,291.8z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M110.2,292.7l-0.3,-2.9h-0.1c-1.3,1.8 -3.8,3.4 -7.1,3.4c-4.7,0 -7.1,-3.3 -7.1,-6.7c0,-5.6 5,-8.7 13.9,-8.6v-0.5c0,-1.9 -0.5,-5.4 -5.3,-5.4c-2.2,0 -4.4,0.7 -6,1.7l-1,-2.8c1.9,-1.2 4.7,-2.1 7.6,-2.1c7.1,0 8.8,4.8 8.8,9.5v8.7c0,2 0.1,4 0.4,5.5H110.2zM109.5,280.9c-4.6,-0.1 -9.8,0.7 -9.8,5.2c0,2.7 1.8,4 4,4c3,0 4.9,-1.9 5.6,-3.9c0.1,-0.4 0.2,-0.9 0.2,-1.3V280.9z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M140.4,258.7v28c0,2.1 0,4.4 0.2,6h-3.8l-0.2,-4h-0.1c-1.3,2.6 -4.1,4.5 -7.9,4.5c-5.6,0 -9.9,-4.7 -9.9,-11.8c0,-7.7 4.7,-12.5 10.4,-12.5c3.5,0 5.9,1.7 7,3.5h0.1v-13.8H140.4zM136.2,278.9c0,-0.5 0,-1.2 -0.2,-1.8c-0.6,-2.7 -2.9,-4.9 -6.1,-4.9c-4.4,0 -6.9,3.8 -6.9,9c0,4.7 2.3,8.6 6.8,8.6c2.8,0 5.4,-1.9 6.2,-5c0.1,-0.6 0.2,-1.2 0.2,-1.8V278.9z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M149.7,281.9c0.1,5.7 3.7,8 8,8c3,0 4.8,-0.5 6.4,-1.2l0.7,3c-1.5,0.7 -4,1.4 -7.7,1.4c-7.1,0 -11.4,-4.7 -11.4,-11.7c0,-7 4.1,-12.5 10.9,-12.5c7.6,0 9.6,6.7 9.6,10.9c0,0.9 -0.1,1.5 -0.1,2H149.7zM162.1,278.8c0.1,-2.7 -1.1,-6.8 -5.8,-6.8c-4.3,0 -6.1,3.9 -6.5,6.8H162.1z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M171.4,275.8c0,-2.4 0,-4.4 -0.2,-6.3h3.7l0.2,3.7h0.1c1.3,-2.2 3.4,-4.3 7.3,-4.3c3.2,0 5.6,1.9 6.6,4.6h0.1c0.7,-1.3 1.6,-2.3 2.6,-3c1.4,-1 2.9,-1.6 5.1,-1.6c3.1,0 7.6,2 7.6,10.1v13.6h-4.1v-13.1c0,-4.5 -1.6,-7.1 -5,-7.1c-2.4,0 -4.3,1.8 -5,3.8c-0.2,0.6 -0.3,1.3 -0.3,2.1v14.3h-4.1v-13.9c0,-3.7 -1.6,-6.4 -4.8,-6.4c-2.6,0 -4.5,2.1 -5.2,4.2c-0.2,0.6 -0.3,1.3 -0.3,2.1v14h-4.1V275.8z"/>
+    <path android:fillColor="#2FAAE1" android:pathData="M212.2,269.5l5.1,13.7c0.5,1.5 1.1,3.4 1.5,4.7h0.1c0.4,-1.4 0.9,-3.2 1.5,-4.8l4.6,-13.6h4.5l-6.3,16.5c-3,7.9 -5.1,12 -7.9,14.5c-2.1,1.8 -4.1,2.5 -5.2,2.7l-1.1,-3.5c1.1,-0.3 2.4,-1 3.7,-2.1c1.1,-0.9 2.6,-2.5 3.5,-4.7c0.2,-0.4 0.3,-0.8 0.3,-1c0,-0.2 -0.1,-0.6 -0.3,-1.1l-8.6,-21.4H212.2z"/>
+    <path android:fillColor="#36A9E1" android:pathData="M17.6,163.1c0.3,0.9 0.6,1.3 0.6,1.3L17.6,163.1z"/>
+    <path android:fillColor="#36A9E1" android:pathData="M211,30c0,0 -5,-13.2 -47.9,-9.4c-43.9,3.9 -81.1,22.6 -103,35.8c-21.8,13.2 -44,31.3 -56.9,59.9C0,123.5 5,134.8 5,134.8l12.6,28.3c-2.3,-5.7 -7.8,-30.1 46.1,-69.3c8,-5.8 18.7,-13.2 51,-27.5c3,-1.3 6,-2.6 9,-3.7l-0.1,168.3h41.9l-0.1,-179.2c11.4,-1.7 19.9,-1.8 23.5,-1.8c25.9,0 35.1,10.5 38,15.7L211,30z"/>
+</vector>
