+++ app/src/main/res/values/colors.xml
@@ -6,5 +6,6 @@
     <color name="text_color_textView1">#575252</color>
     <color name="background_textView2">#F44336</color>
     <color name="text_color_textView2">#673AB7</color>
+    <color name="color_button">#36A9E1</color>
 
 </resources>
