+++ app/src/main/res/values/strings.xml
@@ -3,4 +3,6 @@
     <string name="textView_1">Android</string>
     <string name="textView_2">Java</string>
     <string name="button_change">Поменять</string>
+    <string name="dz1">Dz1</string>
+    <string name="dz2">Dz2</string>
 </resources>
