+++ app/src/main/res/values/styles.xml
@@ -2,7 +2,6 @@
 
     <!-- Base application theme. -->
     <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
-
     </style>
 
 </resources>
